+++ tests/app/btc.spec.js
@@ -5,34 +5,22 @@ const jwt = require('jsonwebtoken');
 const app = require('../../app');
 const { coinlayerRatesService } = require('../../configs/services-config');
 const { HttpCodes } = require('../../helpers/constants');
+const { payload, testRates } = require('./test-data');
 
 jest.mock('../../configs/services-config');
 
 describe('btcRate endpoint', () => {
   describe('GET /btcRate', () => {
-    const testUser = {
-      id: 'd40ddf50-386e-4d6c-a10a-2c08d599ab19',
-      name: 'Software Engineering School',
-      email: 'software@engineering.school',
-    };
-    const { JWT_SECRET_KEY } = process.env;
-    const token = jwt.sign(testUser, JWT_SECRET_KEY);
-    const testRates = {
-      target: 'UAH',
-      rates: {
-        BTC: 1234,
-      },
-    };
+    let token;
 
-    coinlayerRatesService.getBtcToUahRate = jest.fn(() => testRates);
+    beforeAll(() => {
+      const { JWT_SECRET_KEY } = process.env;
+      token = jwt.sign(payload, JWT_SECRET_KEY);
 
-    it('should respond with 401 status code if user is unauthorized', async () => {
-      const response = await request(app).get('/btcRate');
-
-      expect(response.statusCode).toBe(HttpCodes.UNAUTHORIZED);
+      coinlayerRatesService.getBtcToUahRate = jest.fn(() => testRates);
     });
 
-    it('should respond with 200 status code and rates info', async () => {
+    it('should respond with 200 status code and rates info when success', async () => {
       const response = await request(app)
         .get('/btcRate')
         .set('Authorization', `Bearer ${token}`);
@@ -55,6 +43,12 @@ describe('btcRate endpoint', () => {
       } catch (error) {
         expect(next).toHaveBeenCalledWith(error);
       }
+    });
+
+    it('should respond with 401 status code when user is unauthorized', async () => {
+      const response = await request(app).get('/btcRate');
+
+      expect(response.statusCode).toBe(HttpCodes.UNAUTHORIZED);
     });
   });
 });
